Fix the int double error

+++ lib/models/product.dart
@@ -31,7 +31,7 @@ class Product {
     description = json['description'];
     price = json['price'];
     discountPercentage = json['discountPercentage'];
-    rating = json['rating'];
+    rating = json['rating'].toDouble();
     stock = json['stock'];
     brand = json['brand'];
     category = json['category'];
